docs: Fix prioritizeBuilders() example when functions return None
b.getNewestCompleteTime() and b.getOldestRequestTime() can return None
if there are no valid buildrequests. But async_sort() is not able to
sort between None and datetime object. This is fixed by forcing the
minimal possible date when None is returned.

Note that the default sorter in _defaultSorter() has similar code:

   time = yield bldr.getOldestRequestTime()
   if time is None:
       # for builders that do not have pending buildrequest, we just use large number
       time = math.inf
   else:
       if isinstance(time, datetime):
           time = time.timestamp()