common: Support downloading non-latest release artifacts
When creating a release with backports to an old branch one does not
want to click "This is latest release" in the Github GUI, because this
release is not actually the newest one. The releasing code underlying
`make finishrelease` unfortunately only supports retrieving the latest
release. This may lead to wrong files being uploaded to Pypi.

This issue can be fixed by looking into the tag pointing to the current
commit and then looking up a release with matching name.