botmaster: fix access to builder with no config
`AttributeError: 'NoneType' object has no attribute 'workernames'`

On reconfig with new builder(s),
between `BotMaster.reconfigServiceBuilders` and `Builder.reconfigServiceWithBuildbotConfig`,
a builder has it's `config` to `None`.

If a Build finish in this interval, it will eventually call `BotMaster.getBuildersForWorker`
from `Build.buildFinished`, which will throw this exception, preventing the build to be marked
as finished.