buildbot_worker/runprocess.py: Use windows Job API
If process dies, there can still be sub-processes which can not be
killed (on windows platform). This can cause problems with tests running
on buildbot.

Attach spawned process to `JobObject` so if process dies, we can still
kill all its child processes. Now we do not have to kill process (which
does not have to exists) but we terminate whole JobObject, where are all
processes of the main process (included).