www: fix ansi color reset in log view
If we just return cssClasses, we are returning the previous
instance which may be filled with existing colors. That means
a reset sequence `CSI_PREFIX[m`, which has zero ansi classes,
will result in the previous class set being used, rather than
resetting the colors. This is in contrast to `CSI_PREFIX[0m`,
which correctly empties the class set.

Co-authored-by: Monika Kairaityte <monika@kibit.lt>