e2e: Disable WebKit runner due to internal bug in WebKit
WebKit contains a race condition that occurs under certain conditions.

https://bugs.webkit.org/show_bug.cgi?id=280366

It so happens that Buildbot internal CI which runs under gvisor manages
to reproduce the issue quite often.

There exists a pull request that should address the issue.

https://github.com/WebKit/WebKit/pull/34268

For the time being the browser is disabled. It can be re-enabled once
the fix lands into a version of WebKit distributed by Playwright.