Revert "move contrib into buildbot-contrib repository"
This reverts commit 64d835038a936761716e79f43eecfc8d5b71a6ab.

It's best to maintain all functionality related to Buildbot in a single
repository, as then it's easier to keep contrib functionality in sync
with Buildbot code.

For the moment the master/contrib and worker/contrib contents are in the
same state as of beginning 2017 and will need changes transferred from
the buildbot-contrib repository.