changes: Add explicit messages for any HTTP errors during Gerrit polling
This makes the error much less confusing. The previous code attempted to
parse the HTTP body and failed with JSON error. This did not offer any
explanation about what is happening because the HTTP body in case of
e.g. 404 or 400 error message does not need to be descriptive and can be
empty.