db: compressLog spend less time in a DB thread
This is at the cost of having to request action on a DB thread multiple times.
Previous implementation would do (un)compress actions in a DB thread,
which would keep a thread out of the pool for far longer than necessary and
potentially bottlenecking other DB actions (such as users navigating on web UI).