test: No longer run fake DB tests
fakedb is effectively a small in-memory database implementation that
supports only the operations needed by Buildbot. Maintaining this
infrastructure is takes time and is error-prone. A better idea is to use
in-memory sqlite database as the database backend for any test that
needs a database.

The current commit removes the separate fakedb tests, next commits will
convert all other uses.