Ensure consumeErrors=True is passed to gatherResults and DeferredList
Any instance of defer.gatherResults() or defer.DeferredList that are
used with consumeErrors=False is a race condition in tests. If more than
one error is present in the list of deferreds that are waited for, it
will be passed to subsequent callbacks/errbacks attached to the
deferred. If there are no callbacks or errbacks, then an error will be
sent to the logs once the deferred is garbage collected. This may happen
at any time and in many cases will cause a subsequent test to fail.

The error results of the second and subsequent failing deferreds is not
needed. defer.gatherResults() or defer.DeferredList will include the
error result of the first failing deferred. This will still present an
error message for the user to address.

Accordingly, setting consumeErrors=True has few downsides yet it solves
the race conditions in tests.