BuildBot database tests with MariaDB Server
At MariaDB, we've been using BuildBot for several years at buildbot.mariadb.org, with MariaDB Server as the backend database. We would like to ensure that the latest BuildBot code is tested with both the latest stable version and the most recent LTS release of the database server, so we can gain confidence when upgrading BuildBot.

Thank you for your time and for all the great work over the years.