cidb: MariaDB/MySQL healthchecks
{mariadb/mysql}admin ping has the property
that it test the liveness over a unix socket.

Both MySQL and MariaDB containers in the initialization
phases are running with --skip-networking to bootstrap
the database files. As such a healthy check might result
just before the database is restarted without skip-networking.

MariaDB has a healthcheck.sh to encapsuate the logic.

MySQL we use the protocol tcp option to force a connection
attempt on the tcp socket which won't exist during
--skip-networking.