reporters: words: fix build status reporting
I've noticed, that on web interface, build failure status is more verbose, then
on IRC, where it basically just says "failed" and one has to click on the build
to see the reason.

 < bot> Build [#24](https://url/images/#/builders/44/builds/24) of `build` failed.

This patch fixes that by using build.get('state_string') as a fallback instead
of build.get('status_string') in case of build failure.

  <bot> Build [#27](https://url/images/#/builders/119/builds/27) of `build` failed:  failed Images built and installed (failure) (timed out)

Signed-off-by: Petr Štetiar <ynezz@true.cz>