fake: Pass real connector to FakeUpdates
This will be used to pass updates to real connector.

Currently fake data API implementation is not optimal, because it does
not pass data updates to the real data API. The fake data API is used by
fake master. Consequently most data API tests do not exercise the update
data path of the real data API, leading to bugs.

The reason why fake API was implemented this way was because it was
decided that tests should have a quick and dirty access to the effects
of data API updates. Since all tests now use real database one way or
another, it is possible to simply take the data from the database
itself, removing the data API fakeness and improving test coverage.