data: Fix key names in <...>/commits_common_parent Data API return value
If returned ranges are considered as going <from commit> towards <to
commit>, then `from1` and `from2` are incorrect names, because
`commitid1` and `commitid2` in the API path refers to destination
commits, not source commits. In both cases, the common commit is the
source commit.

As a result, it's more appropriate to rename `from1` and `from2`, to
`to1` and `to2` names.