monkeypatches: Work around old workers not being able to connect
The cause of the error is this new code in Twisted:

https://github.com/twisted/twisted/commit/2ac17b0efd15f828214bb54ac5dee8a99b4df75e#diff-3ee4b290d7221a2e082e7c4c040b64321158a7fdb26bfdb5aefebe4c54075e78R488-R492

Old workers may not send parents attribute in their state, thus they
will fail to connect. The same problem may happen on newer workers as
well, but it will only be triggered when an exception is sent to the
master.