github: Do not group dependabot updates
Dependabot grouping is great because it allows to reduce the number of
PRs by grouping small dependency upgrades. It is completely useless
because it merges all upgrades into single commit. As a result, it won't
be possible to easily bisect if problems later appear.

This is the issue in dependabot that tracks this lack of functionality:
https://github.com/dependabot/dependabot-core/issues/8450

This commit disables grouping. There is a script to create a single PR
from multiple dependabot PRs - common/gather_dependabot.py. This will
reduce the overhead of one PR per dependency.