process: allow to collapse build requests when the builder object is not defined.
On asymmetric multi-master configurations where one master runs the Web UI and the
other master works with the builders and workers, it is possible to configure the
Web UI master with an empty list of builders and workers.

However, the code to collapse the build requests runs on the Web UI master and if
this list is empty then it can't find the builder object to evaluate if a custom
collapse function is defined for it, so it fails to collapse anything.

Instead of doing that, it should continue and use the globally defined collapse
function or the default collapse function.

This patch moves the code to evaluate getCollapseRequestsFn to the class
BuildRequestCollapser allowing it to work even when the builder object
is not defined.

This issue was observed on the WebKit's Buildbot CI deployment and there it was
work-arounded by defining also the builders/workers for the Web UI master.
See https://bugs.webkit.org/show_bug.cgi?id=300920